Don't over-use unsigned types since Eigen doesn't support them anyway.

+++ src/circus.cpp
@@ -32,9 +32,9 @@ Eigen::MatrixXf nearest_orthonormal_sinogram(
 {
         // Detect the offset of each column to the sinogram center
         assert(input.rows() > 0 && input.cols() > 0);
-        size_t sinogram_center = (size_t) std::floor((input.rows() - 1) / 2.0);
+        int sinogram_center =  std::floor((input.rows() - 1) / 2.0);
         std::vector<int> offset(input.cols());  // TODO: Eigen vector
-        for (size_t p = 0; p < input.cols(); p++) {
+        for (int p = 0; p < input.cols(); p++) {
                 size_t median = findWeighedMedian(
                         input.data() + p*input.rows(),
                         input.rows());
@@ -45,12 +45,12 @@ Eigen::MatrixXf nearest_orthonormal_sinogram(
         int min = *(std::min_element(offset.begin(), offset.end()));
         int max = *(std::max_element(offset.begin(), offset.end()));
         assert(sgn(min) != sgn(max));
-        size_t padding = (size_t) (std::abs(max) + std::abs(min));
+        int padding = (int) (std::abs(max) + std::abs(min));
         new_center = sinogram_center + max;
         // TODO: zeros?
         Eigen::MatrixXf aligned(input.rows() + padding, input.cols());
-        for (size_t col = 0; col < input.cols(); col++) {
-                for (size_t row = 0; row < input.rows(); row++) {
+        for (int col = 0; col < input.cols(); col++) {
+                for (int row = 0; row < input.rows(); row++) {
                         aligned(max+row-offset[col], col) = input(row, col);
                 }
         }
@@ -72,7 +72,7 @@ Eigen::VectorXf getCircusFunction(
         Eigen::VectorXf output(input.cols());
 
         // Trace all columns
-        for (size_t p = 0; p < input.cols(); p++) {
+        for (int p = 0; p < input.cols(); p++) {
                 float *data = (float*) (input.data() + p*input.rows());
                 size_t length = input.rows();
                 float result;
